requirements.txt
Django==2.2.11
psycopg2==2.7.5
psycopg2-binary==2.7.5
djangorestframework==3.11.0
django-cors-headers==3.2.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
